chore: move transformer tests in to the transformer test file
